+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `name` property `mwc-textfield` & `mwc-textarea` for browser autofill.
 - `ListItem.multipleGraphics` list-item graphic width now configuratble for multiple graphics
 - `Menu.menuCorner` can now configure from which horizontal corner should the menu anchor from.
+- Add `touch` param to `mwc-checkbox` to control touchscreen accessibility.
 
 ### Changed